fix: relocate helper files to plugin and component scope
