+++ packages/component-library/src/plugin/device-helper.plugin.ts
@@ -1,6 +1,5 @@
 import type { Plugin } from "vue";
-// @ts-expect-error
-import DeviceHelper from "../helper/device.helper";
+import DeviceHelper from "./device-helper.service";
 
 const DeviceHelperPlugin: Plugin = {
   install(app) {
