+++ packages/component-library/src/plugin/device-helper.service.ts
@@ -1,141 +1,118 @@
 import { debounce } from "@/utils/debounce";
 
+interface ResizeListenerConfig {
+  listener: (event: Event) => void;
+  scope?: unknown;
+  component?: unknown;
+}
+
 /**
  * The DeviceHelper provides methods to get device and browser information like the current viewport size.
  * The helper methods can be accessed with "this.$device" in every Vue component.
- *
- * @constructor
  */
-function DeviceHelper() {
-  this.listeners = [];
+export default class DeviceHelper {
+  private listeners: ResizeListenerConfig[] = [];
 
-  window.addEventListener("resize", this.resize.bind(this));
-}
-
-DeviceHelper.prototype = Object.assign(DeviceHelper.prototype, {
   /**
    * Resize method which will be fired when the user resizes the browser.
-   *
-   * @returns {void}
    */
-  resize: debounce(function debouncedResize(event) {
+  resize = debounce((event: Event): void => {
     this.listeners.forEach((listenerObject) => {
       listenerObject.listener.call(listenerObject.scope, event);
     });
-  }, 100),
+  }, 100);
+
+  constructor() {
+    window.addEventListener("resize", this.resize.bind(this));
+  }
 
   /**
    * Registers an event register for the browser "resize" event.
-   *
-   * @param {Function} callback
-   * @param {Any} scope
-   * @param {Object} component
-   * @returns {number}
    */
-  onResize({ listener, scope, component }) {
+  onResize({ listener, scope, component }: ResizeListenerConfig): number {
     if (!scope) {
       scope = window;
     }
+
     this.listeners.push({ listener, scope, component });
     return this.listeners.length - 1;
-  },
+  }
 
-  removeResizeListener(component) {
+  removeResizeListener(component: unknown): boolean {
     this.listeners = this.listeners.filter(
       (listenerObject) => component !== listenerObject.component,
     );
 
     return true;
-  },
+  }
 
   /**
    * Returns the user agent string.
-   *
-   * @returns {string}
    */
-  getUserAgent() {
+  getUserAgent(): string {
     return window.navigator.userAgent;
-  },
+  }
 
   /**
-   * Returns the current viewport with in pixels.
-   * @returns {number}
+   * Returns the current viewport width in pixels.
    */
-  getViewportWidth() {
+  getViewportWidth(): number {
     return window.innerWidth;
-  },
+  }
 
   /**
    * Returns the current viewport height in pixels.
-   *
-   * @returns {number}
    */
-  getViewportHeight() {
+  getViewportHeight(): number {
     return window.innerHeight;
-  },
+  }
 
   /**
    * Returns the pixel ratio of the device as a number.
-   *
-   * @returns {number}
    */
-  getDevicePixelRatio() {
+  getDevicePixelRatio(): number {
     return window.devicePixelRatio;
-  },
+  }
 
   /**
    * Returns the device screen width in pixels.
-   *
-   * @returns {number}
    */
-  getScreenWidth() {
+  getScreenWidth(): number {
     return window.screen.width;
-  },
+  }
 
   /**
    * Returns the device screen height in pixels.
-   *
-   * @returns {number}
    */
-  getScreenHeight() {
+  getScreenHeight(): number {
     return window.screen.height;
-  },
+  }
 
   /**
    * Returns information about the screen orientation.
-   *
-   * @returns {object}
    */
-  getScreenOrientation() {
+  getScreenOrientation(): ScreenOrientation {
     return window.screen.orientation;
-  },
+  }
 
   /**
    * Returns the current browser language as a string.
-   *
-   * @returns {string}
    */
-  getBrowserLanguage() {
+  getBrowserLanguage(): string {
     return window.navigator.language;
-  },
+  }
 
   /**
    * Returns the current platform (e.g. "Win32") as a string.
-   *
-   * @returns {string}
    */
-  getPlatform() {
+  getPlatform(): string {
     return window.navigator.platform;
-  },
+  }
 
   /**
    * Returns the system-key (e.g. "CTRL") as a string depending of the current operating system.
-   *
-   * @returns {string}
    */
-  getSystemKey() {
+  getSystemKey(): string {
     return this.getPlatform().indexOf("Mac") > -1 ? "CTRL" : "ALT";
-  },
-});
-
-export default DeviceHelper;
+  }
+}
